requirements.txt
# ===== 60天冲刺项目核心依赖 =====

# Web前端
streamlit>=1.32.0

# 原有依赖
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.7.0
asttokens==2.4.1
attrs==24.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
chromadb==1.0.16
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
curl_cffi==0.11.2
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==5.5
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
durationpy==0.10
executing==2.1.0
fastapi==0.115.6
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
finnhub-python==2.4.24
flatbuffers==25.2.10
fonttools==4.55.3
fredapi==0.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
gradio==5.12.0
gradio_client==1.5.4
greenlet==3.2.4
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kubernetes==33.1.0
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.0.10
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.19.1
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.99.9
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.12
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==10.4.0
platformdirs==4.3.6
posthog==5.4.0
praw==7.8.1
prawcore==2.4.0
prettytable==3.16.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.25.0
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydub==0.25.1
pyecharts==2.0.8
Pygments==2.18.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.1.0
referencing==0.35.1
regex==2025.7.34
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
safehttpx==0.1.6
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
simfin==1.0.1
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.41.3
stockstats==0.6.5
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.61
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
